+++ man-roxygen/dependencies.R
@@ -2,4 +2,4 @@
 #'
 #' * `.pdf` output requires a full LaTeX installation on the local computer.
 #' * `.png` output requires the `webshot2` package.
-#' * `.html` self-contained files require the `base64enc` package.
+#' * `.html` self-contained files and `theme_typst(portable = TRUE)` require the `base64enc` package.